Rewrite Daphne integration test
