+++ .github/workflows/ci-build.yml
@@ -98,7 +98,6 @@ jobs:
     env:
       CARGO_INCREMENTAL: 0
       CARGO_TERM_COLOR: always
-      DAPHNE_INTEROP_CONTAINER: skip
       JANUS_INTEROP_CONTAINER: skip
       RUSTDOCFLAGS: "-D warnings"
       RUSTFLAGS: "-D warnings"
