+++ integration_tests/src/lib.rs
@@ -1,6 +1,6 @@
 //! This crate contains functionality useful for Janus integration tests.
 
 pub mod client;
-#[cfg(feature = "daphne")]
 pub mod daphne;
+pub mod interop_api;
 pub mod janus;
