+++ integration_tests/tests/daphne.rs
@@ -1,36 +1,32 @@
-#![cfg(feature = "daphne")]
-
 use common::{submit_measurements_and_verify_aggregate, test_task_builders};
-use janus_aggregator_core::task::Task;
+use janus_aggregator_core::task::{QueryType, Task};
 use janus_core::{
     task::VdafInstance,
     test_util::{install_test_trace_subscriber, testcontainers::container_client},
 };
-use janus_integration_tests::{daphne::Daphne, janus::Janus};
+use janus_integration_tests::{client::ClientBackend, daphne::Daphne, janus::Janus};
 use janus_interop_binaries::test_util::generate_network_name;
 use janus_messages::Role;
 
 mod common;
 
-// TODO(timg): re-enable DAphne tests once Daphne implements DAP-04
-
 // This test places Daphne in the leader role & Janus in the helper role.
 #[tokio::test(flavor = "multi_thread")]
-#[ignore]
+#[ignore = "Daphne does not yet publish a leader container image"]
 async fn daphne_janus() {
     install_test_trace_subscriber();
 
     // Start servers.
     let network = generate_network_name();
     let (collector_private_key, leader_task, helper_task) =
-        test_task_builders(VdafInstance::Prio3Count);
+        test_task_builders(VdafInstance::Prio3Count, QueryType::TimeInterval);
 
-    // Daphne is hardcoded to serve from a path starting with /v01/.
+    // Daphne is hardcoded to serve from a path starting with /v04/.
     let [leader_task, helper_task]: [Task; 2] = [leader_task, helper_task]
         .into_iter()
         .map(|task| {
             let mut endpoints = task.aggregator_endpoints().to_vec();
-            endpoints[Role::Leader.index().unwrap()].set_path("/v01/");
+            endpoints[Role::Leader.index().unwrap()].set_path("/v04/");
             task.with_aggregator_endpoints(endpoints).build()
         })
         .collect::<Vec<_>>()
@@ -46,27 +42,27 @@ async fn daphne_janus() {
         (leader.port(), helper.port()),
         &leader_task,
         &collector_private_key,
+        &ClientBackend::InProcess,
     )
     .await;
 }
 
 // This test places Janus in the leader role & Daphne in the helper role.
 #[tokio::test(flavor = "multi_thread")]
-#[ignore]
 async fn janus_daphne() {
     install_test_trace_subscriber();
 
     // Start servers.
     let network = generate_network_name();
     let (collector_private_key, leader_task, helper_task) =
-        test_task_builders(VdafInstance::Prio3Count);
+        test_task_builders(VdafInstance::Prio3Count, QueryType::TimeInterval);
 
-    // Daphne is hardcoded to serve from a path starting with /v01/.
+    // Daphne is hardcoded to serve from a path starting with /v04/.
     let [leader_task, helper_task]: [Task; 2] = [leader_task, helper_task]
         .into_iter()
         .map(|task| {
             let mut endpoints = task.aggregator_endpoints().to_vec();
-            endpoints[Role::Helper.index().unwrap()].set_path("/v01/");
+            endpoints[Role::Helper.index().unwrap()].set_path("/v04/");
             task.with_aggregator_endpoints(endpoints).build()
         })
         .collect::<Vec<_>>()
@@ -82,6 +78,7 @@ async fn janus_daphne() {
         (leader.port(), helper.port()),
         &leader_task,
         &collector_private_key,
+        &ClientBackend::InProcess,
     )
     .await;
 }
